# openapi.yaml
openapi: 3.0.3
info:
  title: LoadFlow API
  version: 1.3.0
  description: |
    Welcome to the LoadFlow API!
    
    This API provides a complete backend for the LoadFlow Dispatch Application, allowing dispatchers, drivers, and administrators to orchestrate logistics seamlessly.

    ### 🔑 Authentication

    The system uses **Cookie-based JWT Authentication**.

    1. **Login** or **Register** (if invited) to receive a secure HTTP-only cookie.
    2. This cookie is automatically sent with subsequent requests.
    3. **Note for API Tools (Postman/Swagger):** Ensure cookies are enabled/stored after login so protected endpoints can be tested.

    ### 📦 Primary Features
    
    This API is built explicitly to handle fleet operations.
    
    - **Logistics Data:** Access resources in JSON format.
    - **Driver Dashboard Metrics:** Real-time calculation of Service Hours, Miles, and Completed Missions.
    - **Smart Notifications:** Optimistic state updates and ultra-fast polling for instant driver/dispatcher feedback.
    - **Proof of Delivery:** Secure `multipart/form-data` uploads handled and routed towards Cloudinary.
    - **Server Actions:** Public gateways directly dispatch SMTP NodeMailer notifications without persistence overhead.

    ### 👥 User Roles

    - **Driver:** Can view assigned configurations, update transit statuses, track performance metrics (Miles/Hours), and submit Proof of Delivery.
...
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login and mint the auth cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded
          headers:
            Set-Cookie:
              schema:
                type: string
              description: HttpOnly auth cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginSuccessResponse'
        '400':
          description: Missing email or password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LockedAccountResponse'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the currently authenticated user
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: No valid auth cookie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password after admin approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Validation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Reset not approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/profile/update:
    post:
      tags: [Auth]
      summary: Update the logged-in user's profile and optionally change password
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    type: object
                    properties:
                      name:
                        type: string
                      role:
                        type: string
                        enum: [Admin, Dispatcher, Driver]
                      email:
                        type: string
                        format: email
                      city:
                        type: string
                        nullable: true
                      province:
                        type: string
                        nullable: true
        '400':
          description: Invalid update request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized or current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/driver-requests:
    post:
      tags: [Fleet Requests]
      summary: Submit a driver request
      description: Dispatcher only. Notifies Admins of a new driver candidate.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [firstName, lastName, email]
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Request sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Validation error
        '403':
          description: Forbidden
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  schemas:
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    User:
      type: object
      properties:
        _id:
          type: string
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [Admin, Dispatcher, Driver]
        isPending:
          type: boolean
        phone:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        dob:
          type: string
          format: date-time
          nullable: true
        city:
          type: string
          nullable: true
        province:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        resetPasswordRequested:
          type: boolean
        resetPasswordApproved:
          type: boolean
        loginAttempts:
          type: integer
        isLocked:
          type: boolean
        lockedUntil:
          type: string
          format: date-time
          nullable: true
        tokenVersion:
          type: integer
          description: Session revocation index
        isTwoFactorEnabled:
          type: boolean
          description: Whether the user has setup 2FA
        requestedBy:
          type: string
          nullable: true
          description: Dispatcher ID for recruitment flow reference
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        __v:
          type: integer
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginSuccessResponse:
      type: object
      properties:
        message:
          type: string
        user:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            role:
              type: string
              enum: [Admin, Dispatcher, Driver]
    LockedAccountResponse:
      type: object
      properties:
        error:
          type: string
        locked:
          type: boolean
        message:
          type: string
...
    ResetPasswordRequest:
      type: object
      required: [email, password, confirmPassword]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        confirmPassword:
          type: string
    ProfileUpdateRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        licenseNumber:
          type: string
        currentPassword:
          type: string
        newPassword:
          type: string
...